move Asset to it's own interface file
