+++ types/interfaces/AssetError.ts
@@ -1,0 +1,4 @@
+export declare interface AssetError {
+  type: 'invalid_input' | string;
+  messages: string[];
+}
