+++ types/interfaces/AssetMaster.ts
@@ -1,0 +1,4 @@
+export declare interface AssetMaster {
+  status: 'ready' | 'preparing' | 'errored';
+  url: string;
+}
