+++ types/interfaces/Track.ts
@@ -1,0 +1,5 @@
+import { AudioTrack } from './AudioTrack';
+import { TextTrack } from './TextTrack';
+import { VideoTrack } from './VideoTrack';
+
+export declare type Track = VideoTrack | AudioTrack | TextTrack;
